feat: New Footer & rename Old Footer to LegacyFooter
chore: Rename Old Footer to LegacyFooter and move the component to its directory

remove unused styles in the storybook style.css

fix: remove div wrapped in a footer element, reuse ablyawards from menganav, simplify loop
