+++ .storybook/styles.css
@@ -8,7 +8,7 @@
 @import "../src/core/Code/component.css";
 @import "../src/core/CookieMessage/component.css";
 @import "../src/core/Flash/component.css";
-@import "../src/core/Footer/component.css";
+@import "../src/core/LegacyFooter/component.css";
 @import "../src/core/LegacyMeganav/component.css";
 @import "../src/core/Notice/component.css";
 @import "../src/core/Slider/component.css";
