+++ src/core/LegacyFooter/component.css
